+++ amulet/api/level/immutable_structure/void_format_wrapper.py
