+++ tests/test_loader.py
@@ -3,7 +3,7 @@ import unittest
 from test_utils import get_world_path
 
 from amulet.world_interface import load_world
-from amulet.api import world
+from amulet.api import level
 
 
 @unittest.skip
@@ -23,11 +23,11 @@ class DefinitionBasedLoaderTestCase(unittest.TestCase):
     def test_loading(self):
         # world_obj = self.loader.load_world(get_world_path("1.12.2 World"))
         world_obj = load_world(get_world_path("1.12.2 World"))
-        self.assertIsInstance(world_obj, world.World)
+        self.assertIsInstance(world_obj, level.World)
 
         # world_obj = self.loader.load_world(get_world_path("1.13 World"))
         world_obj = load_world(get_world_path("1.13 World"))
-        self.assertIsInstance(world_obj, world.World)
+        self.assertIsInstance(world_obj, level.World)
 
 
 if __name__ == "__main__":
